refactor: graphic commandline split
